// packages/angular/src/app/client/service.model.ts
export class Service {
  _id: string;
  client: string;
  ipAddress: string;
  servicePlan: string;
  dateFrom: string;
  closeDate: string;
  userName: string;
  password: string;
  note: string;
  isActive: boolean;
  payment: string;
  user: string;
  accessPoint: string;
  macAddress: string;
  lastOutage: string;
  createdAt: string;
}